Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-cbc-import-key-unwrap-jwk-rsa-key-private-expected.txt

Test unwrapping a JWK RSA private key with AES-CBC using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS unwrappedKey.kty is jwkKey.kty
PASS unwrappedKey.alg is jwkKey.alg
PASS unwrappedKey.key_ops is jwkKey.key_ops
PASS unwrappedKey.ext is jwkKey.ext
PASS unwrappedKey.n is jwkKey.n
PASS unwrappedKey.e is jwkKey.e
PASS unwrappedKey.p is jwkKey.p
PASS unwrappedKey.q is jwkKey.q
PASS unwrappedKey.dp is jwkKey.dp
PASS unwrappedKey.dq is jwkKey.dq
PASS unwrappedKey.qi is jwkKey.qi
PASS successfullyParsed is true

TEST COMPLETE
